(minor) extract sub-components from the parent components for HomeDocumentationSection, HomeExamplesSection, HomeFeaturesSection, and HomeHero, and remove white space in HomeFeaturesSection
