+++ src/components/Home/HomeDocumentationSection/Card.js
@@ -1,0 +1,34 @@
+import { useResponsive } from 'hooks/useResponsive'
+import { Box, Heading, Paragraph } from 'grommet'
+import { Column } from 'components/shared/Column'
+
+export const Card = ({ heading, body, footer, img, ...props }) => {
+  const { setResponsive } = useResponsive()
+
+  return (
+    <Column
+      background={{
+        image: `url('${img}')`,
+        position: 'bottom right',
+        repeat: 'no-repeat',
+        size: '100%'
+      }}
+      elevation="medium"
+      pad={{
+        horizontal: setResponsive('large', 'basex6', 'basex7'),
+        top: setResponsive('large', 'xlarge'),
+        bottom: setResponsive('large', 'xlarge', 'basex8')
+      }}
+      // eslint-disable-next-line react/jsx-props-no-spreading
+      {...props}
+    >
+      <Heading level={1} margin={{ bottom: setResponsive('medium', 'small') }}>
+        {heading}
+      </Heading>
+      <Paragraph>{body}</Paragraph>
+      <Box margin={{ top: setResponsive('large', 'medium') }}>{footer}</Box>
+    </Column>
+  )
+}
+
+export default Card
